actions配下の処理をmemoApp.jsに統合
