--- 04.class/actions/deleteMemo.js
@@ -1,12 +1,0 @@
-import { DELETE_MEMO_BY_ID } from "../db/queries.js";
-import { DELETE_MEMO_LOG_MESSAGES } from "../config/log.js";
-import { handleMemoAction } from "../helpers/memoHelpers.js";
-
-export async function deleteMemo(db) {
-  await handleMemoAction(
-    db,
-    DELETE_MEMO_LOG_MESSAGES.PROMPT,
-    DELETE_MEMO_BY_ID,
-    () => DELETE_MEMO_LOG_MESSAGES.SUCCESS,
-  );
-}
