--- 04.class/actions/readMemo.js
@@ -1,12 +1,0 @@
-import { SELECT_MEMO_BY_ID } from "../db/queries.js";
-import { READ_MEMO_LOG_MESSAGES } from "../config/log.js";
-import { handleMemoAction } from "../helpers/memoHelpers.js";
-
-export async function readMemo(db) {
-  await handleMemoAction(
-    db,
-    READ_MEMO_LOG_MESSAGES.PROMPT,
-    SELECT_MEMO_BY_ID,
-    (row) => `\n${row.memo}\n`,
-  );
-}
